+++ src/main/kotlin/com/cmgapps/intellij/ProguardRetraceUnscrambler.kt
@@ -16,38 +16,23 @@
 
 package com.cmgapps.intellij
 
-import com.android.tools.r8.Diagnostic
-import com.android.tools.r8.DiagnosticsHandler
-import com.android.tools.r8.retrace.ProguardMapProducer
-import com.android.tools.r8.retrace.ProguardMappingSupplier
-import com.android.tools.r8.retrace.RetraceStackTraceContext
-import com.android.tools.r8.retrace.StringRetrace
+import com.intellij.icons.AllIcons
 import com.intellij.ide.util.PropertiesComponent
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.ui.Messages
 import com.intellij.unscramble.UnscrambleSupport
-import org.jetbrains.annotations.VisibleForTesting
-import proguard.retrace.ReTrace
-import java.awt.BorderLayout
+import java.awt.Component
 import java.awt.Dimension
 import java.awt.event.ItemEvent
-import java.io.ByteArrayOutputStream
+import java.awt.event.MouseAdapter
+import java.awt.event.MouseEvent
 import java.io.File
-import java.io.LineNumberReader
-import java.io.PrintStream
-import java.io.PrintWriter
-import java.io.StringWriter
 import java.util.ResourceBundle
-import javax.swing.Action
-import javax.swing.BoxLayout
+import javax.swing.Box
 import javax.swing.JCheckBox
-import javax.swing.JComponent
 import javax.swing.JLabel
-import javax.swing.JPanel
-import javax.swing.SwingConstants
 
 typealias RetracerFactory = (
     mappingFile: File,
@@ -67,24 +52,9 @@ private val DEFAULT_RETRACER_FACTORY: RetracerFactory = { mappingFile, logger, v
     )
 }
 
-typealias ErrorDialogFactory = (
-    project: Project,
-    bundle: ResourceBundle,
-    mappingFile: File,
-) -> DialogWrapper
-
-private val DEFAULT_ERROR_DIALOG_FACTORY: ErrorDialogFactory = { project, bundle, mappingFile ->
-    ErrorDialog(
-        project,
-        bundle,
-        mappingFile.name,
-    )
-}
-
 class ProguardRetraceUnscrambler(
     private val retracerFactory: RetracerFactory = DEFAULT_RETRACER_FACTORY,
-    private val errorDialogFactory: ErrorDialogFactory = DEFAULT_ERROR_DIALOG_FACTORY,
-) : UnscrambleSupport<JPanel> {
+) : UnscrambleSupport<Box> {
     private val bundle = ResourceBundle.getBundle("Bundle")
     private val properties = PropertiesComponent.getInstance()
 
@@ -94,22 +64,28 @@ class ProguardRetraceUnscrambler(
         project: Project,
         text: String,
         logName: String,
-        settings: JPanel?,
+        settings: Box?,
     ): String? {
+        requireNotNull(settings) { "Settings cannot be null" }
+
         if (logName.isBlank() || text.isBlank()) return text
 
         val mappingFile = File(logName)
 
         if (!mappingFile.exists()) {
-            errorDialogFactory(project, bundle, mappingFile).show()
+            Messages.showErrorDialog(project, bundle.getString("error_text").format(logName), "")
             return null
         }
 
-        val useR8: Boolean = (settings?.getComponent(USE_R8_INDEX) as? JCheckBox)?.isSelected ?: false
+        val r8SettingsPanel = (settings.getComponent(USE_R8_INDEX) as Box)
+
+        val useR8: Boolean =
+            (r8SettingsPanel.getComponent(0) as JCheckBox).isSelected
 
         val allClassNamesSetting: Boolean =
-            (settings?.getComponent(ALL_CLASS_NAMES_INDEX) as? JCheckBox)?.isSelected ?: false
-        val verboseSetting: Boolean = (settings?.getComponent(VERBOSE_INDEX) as? JCheckBox)?.isSelected ?: false
+            (settings.getComponent(ALL_CLASS_NAMES_INDEX) as JCheckBox).isSelected
+
+        val verboseSetting: Boolean = (settings.getComponent(VERBOSE_INDEX) as JCheckBox).isSelected
 
         val logger = logger<ProguardRetraceUnscrambler>()
 
@@ -131,44 +107,72 @@ class ProguardRetraceUnscrambler(
         }
     }
 
-    override fun createSettingsComponent() =
-        JPanel().apply {
-            layout = BoxLayout(this, BoxLayout.Y_AXIS)
+    override fun createSettingsComponent(): Box =
+        Box.createVerticalBox().let { container ->
+            container.alignmentX = Component.LEFT_ALIGNMENT
+            val useR8Label = bundle.getString("use_r8")
 
             val useR8 =
-                JCheckBox(bundle.getString("use_r8"))
+                JCheckBox(useR8Label, Messages.getInformationIcon(), properties.getBoolean(USE_R8_PROPERTY, false))
                     .apply {
-                        isSelected = properties.getBoolean(USE_R8_PROPERTY, false)
                         addItemListener {
                             properties.setValue(USE_R8_PROPERTY, it.stateChange == ItemEvent.SELECTED)
                         }
                     }
-            add(useR8, USE_R8_INDEX)
 
-            JCheckBox(bundle.getString("verbose_text"))
+            Box
+                .createHorizontalBox()
                 .apply {
-                    isSelected = properties.getBoolean(VERBOSE_PROPERTY, false)
+                    alignmentX = Component.LEFT_ALIGNMENT
+
+                    add(useR8)
+                    add(Box.createRigidArea(Dimension(5, 0)))
+                    add(
+                        JLabel(
+                            AllIcons.General.Note,
+                        ).apply {
+                            addMouseListener(
+                                object : MouseAdapter() {
+                                    override fun mouseClicked(e: MouseEvent) {
+                                        if (e.button == MouseEvent.BUTTON1) {
+                                            Messages.showMessageDialog(
+                                                bundle.getString("use_r8_note"),
+                                                useR8Label,
+                                                Messages.getInformationIcon(),
+                                            )
+                                        }
+                                    }
+                                },
+                            )
+                        },
+                    )
+                }.let(container::add)
+
+            JCheckBox(bundle.getString("verbose_text"), properties.getBoolean(VERBOSE_PROPERTY, false))
+                .apply {
                     addItemListener {
                         properties.setValue(VERBOSE_PROPERTY, it.stateChange == ItemEvent.SELECTED)
                     }
                 }.let {
-                    add(it, VERBOSE_INDEX)
+                    container.add(it)
                 }
 
             val allClassNames =
-                JCheckBox(bundle.getString("all_class_names_text"))
-                    .apply {
-                        isEnabled = !useR8.isSelected
-                        isSelected = properties.getBoolean(ALL_CLASS_NAMES_PROPERTY, false)
-                        addItemListener {
-                            properties.setValue(ALL_CLASS_NAMES_PROPERTY, it.stateChange == ItemEvent.SELECTED)
-                        }
+                JCheckBox(
+                    bundle.getString("all_class_names_text"),
+                    properties.getBoolean(ALL_CLASS_NAMES_PROPERTY, false),
+                ).apply {
+                    isEnabled = !useR8.isSelected
+                    addItemListener {
+                        properties.setValue(ALL_CLASS_NAMES_PROPERTY, it.stateChange == ItemEvent.SELECTED)
                     }
-            add(allClassNames, ALL_CLASS_NAMES_INDEX)
+                }
+            container.add(allClassNames)
 
             useR8.addItemListener {
                 allClassNames.isEnabled = it.stateChange != ItemEvent.SELECTED
             }
+            container
         }
 
     private companion object {
@@ -179,148 +183,5 @@ class ProguardRetraceUnscrambler(
         private const val ALL_CLASS_NAMES_PROPERTY = "$PROPERTIES_PREFIX.all-class-names"
         private const val VERBOSE_PROPERTY = "$PROPERTIES_PREFIX.verbose"
         private const val USE_R8_PROPERTY = "$PROPERTIES_PREFIX.use_r8"
-    }
-}
-
-class ErrorDialog(
-    project: Project,
-    private val bundle: ResourceBundle,
-    private val fileName: String,
-) : DialogWrapper(project, false) {
-    init {
-        init()
-    }
-
-    @VisibleForTesting
-    public override fun createActions(): Array<Action> = arrayOf(okAction)
-
-    @VisibleForTesting
-    public override fun createCenterPanel(): JComponent =
-        JPanel(BorderLayout()).apply {
-            JLabel(
-                bundle.getString("error_text").format(fileName),
-                Messages.getWarningIcon(),
-                SwingConstants.HORIZONTAL,
-            ).apply {
-                iconTextGap = 10
-                preferredSize = Dimension(300, 100)
-            }.let {
-                add(it, BorderLayout.CENTER)
-            }
-        }
-}
-
-private class R8Retracer(
-    private val retrace: StringRetrace,
-) : Retracer {
-    override fun retraceStackTrace(stacktrace: String): String {
-        val parsedStacktrace =
-            LineNumberReader(stacktrace.reader()).use { reader ->
-                reader.readLines()
-            }
-
-        val writer = StringWriter()
-        writer.use {
-            retrace
-                .retraceStackTrace(parsedStacktrace, RetraceStackTraceContext.empty())
-                .result
-                .forEach { frameResults ->
-                    frameResults.ambiguousResult.forEach {
-                        writer.append(it.result.joinToString(separator = "\n")).append("\n")
-                    }
-                }
-        }
-
-        return writer.toString()
-    }
-
-    companion object {
-        // This is a slight modification of the default regular expression shown for proguard-retrace
-        // that allows for retracing classes in the form <class>: lorem ipsum...
-        // Seems like Proguard retrace is expecting the form "Caused by: <class>".
-        // com/android/tools/r8/retrace/internal/StackTraceRegularExpressionParser.java
-        const val DEFAULT_REGULAR_EXPRESSION =
-            "(?:.*?\\bat\\s+%c\\.%m\\s*\\(%S\\)\\p{Z}*(?:~\\[.*\\])?)" +
-                "|(?:(?:(?:%c|.*)?[:\"]\\s+)?%c(?:(:|]).*)?)"
-    }
-}
-
-private class ProguardRetracer(
-    private val retrace: ReTrace,
-) : Retracer {
-    override fun retraceStackTrace(stacktrace: String): String {
-        val buffer = ByteArrayOutputStream()
-        LineNumberReader(stacktrace.reader().buffered()).use { reader ->
-            PrintWriter(buffer).use { writer ->
-                retrace.retrace(reader, writer)
-            }
-        }
-        return buffer.toString()
-    }
-}
-
-fun interface Retracer {
-    fun retraceStackTrace(stacktrace: String): String
-
-    companion object {
-        fun create(
-            mappingFile: File,
-            logger: Logger,
-            verbose: Boolean,
-            allClassNames: Boolean,
-            useR8: Boolean,
-        ): Retracer {
-            if (useR8) {
-                logger.info("Using R8 Retrace")
-                return R8Retracer(
-                    StringRetrace.create(
-                        ProguardMappingSupplier
-                            .builder()
-                            .setProguardMapProducer(ProguardMapProducer.fromPath(mappingFile.toPath()))
-                            .build(),
-                        R8DiagnosticsHandler(logger),
-                        R8Retracer.DEFAULT_REGULAR_EXPRESSION,
-                        verbose,
-                    ),
-                )
-            }
-
-            logger.info("Using Proguard Retrace")
-
-            return ProguardRetracer(
-                ReTrace(
-                    ReTrace.REGULAR_EXPRESSION,
-                    ReTrace.REGULAR_EXPRESSION2,
-                    allClassNames,
-                    verbose,
-                    mappingFile,
-                ),
-            )
-        }
-    }
-}
-
-internal class R8DiagnosticsHandler(
-    private val logger: Logger,
-) : DiagnosticsHandler {
-    override fun error(error: Diagnostic) {
-        ByteArrayOutputStream().use {
-            DiagnosticsHandler.printDiagnosticToStream(error, "R8 Retrace Error", PrintStream(it))
-            logger.error(it.toString())
-        }
-    }
-
-    override fun warning(error: Diagnostic) {
-        ByteArrayOutputStream().use {
-            DiagnosticsHandler.printDiagnosticToStream(error, "R8 Retrace Warning", PrintStream(it))
-            logger.warn(it.toString())
-        }
-    }
-
-    override fun info(error: Diagnostic) {
-        ByteArrayOutputStream().use {
-            DiagnosticsHandler.printDiagnosticToStream(error, "R8 Retrace Info", PrintStream(it))
-            logger.info(it.toString())
-        }
     }
 }